refactor: extract tick logic into engine
